chore: remove ReactiveDialogLayout
